Fix python names of RSAPublicKey, DSAPublicKey
In 079d4f65a743fb9e952ab109c1a24997c15398a7, I was a little greedy on my
find/replace and ended up adjusting the Python names of these structs as
well as the C names. We wish to keep the Python names the same, so
revert that part of the patch.

Credit to Stanislav Levin for finding and reporting this.

Signed-off-by: Alexander Scheel <ascheel@redhat.com>